requirements.txt
# Core search dependencies
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=1.5.0
openpyxl>=3.0.10
pdfminer.six>=20221105

# AI search dependencies
sentence-transformers>=2.2.2
chromadb>=0.4.15
torch>=1.13.0
transformers>=4.21.0

# Windows-specific dependencies
pywin32>=305; sys_platform == "win32"

# Additional dependencies for converter tools
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfminer.six>=20221105
openpyxl>=3.0.10
pandas>=1.5.0 